requirements.txt
fastapi
uvicorn
pytest
openai
sqlalchemy[asyncio]
asyncpg
alembic
python-dotenv
python-multipart
